+++ src/Packman/Providers/IPackageAsset.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Packman
+{
+    public interface IPackageAsset
+    {
+        IEnumerable<string> Files { get; set; }
+        string MainFile { get; set; }
+        string Version { get; set; }
+    }
+}