+++ src/Packman/Providers/IPackageProvider.cs
@@ -10,6 +10,7 @@ namespace Packman
         Task<InstallablePackage> GetInstallablePackage(string packageName, string version);
         Task<IEnumerable<string>> GetPackageNamesAsync();
         Task<IEnumerable<string>> GetVersionsAsync(string packageName);
+        Task<IPackageMetaData> GetPackageMetaDataAsync(string packageName);
         Task InitializeAsync();
         bool IsInitialized { get; }
     }