+++ CHANGELOG.md
@@ -14,11 +14,11 @@
   - [x] Select exactly which files to install
   - [x] Optional creation of package folders
   - [x] Preselect latest stable version
-- [ ] JSON manifest Intellisense
+- [x] JSON manifest Intellisense
   - [x] JSON Schema
   - [x] Package name Intellisense
   - [x] Package version Intellisense
-  - [ ] Package file array Intellisense
+  - [x] Package file array Intellisense
 - [ ] JSON manifest validation
   - [x] JSON Schema
   - [ ] Package version validation